+++ backend/src/controllers/reports/Tab_50_rpt_PF_NetRecoverySummaryByQuarter.js
@@ -22,9 +22,13 @@ controller.Tab_50_rpt_PF_NetRecoverySummaryByQuarter = async (request, reply) =>
   try {
     // Get the data from the database.
     const getDate = async () => new Date();
+    const report = await model.Tab_50_rpt_PF_NetRecoverySummaryByQuarter(request.query);
+    const report_totals = await model.Tab_50_totals(request.query);
 
     const result = {
       report_date: await getDate(),
+      report,
+      report_totals,
     };
 
     const body = await getDocumentApiBody(result, "Tab_50_rpt_PF_NetRecoverySummaryByQuarter.docx");